+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 #### Added
 - Warn if attaching a kprobe to a non-traceable function
   - [#1835](https://github.com/iovisor/bpftrace/pull/1835)
+- Support for `-k[k]` and `elapsed` in `iter` probes
+  - [#1882](https://github.com/iovisor/bpftrace/pull/1882)
 
 #### Changed
 - Disallow accessing common tracepoint fields